nlp-service/requirements.txt
flask==2.3.3
flask-cors==3.0.10
nltk==3.8.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
vaderSentiment==3.3.2
python-dotenv==1.0.0
gunicorn==20.1.0
google-generativeai==0.3.2
